refactor(lcm): split embeddings.py into embedder/storage/semantic/hybrid package
``backend/app/core/lcm/embeddings.py`` was 548 lines — over the 500-line
file budget. Promote to a ``backend/app/core/lcm/embeddings/`` package
per the project's ``package-multi-file-features`` rule. Four role-named
modules + a re-exporting ``__init__.py`` keep the public surface
identical for every existing call site.

Module layout:
- ``embedder.py`` — ``Embedder`` Protocol, ``DeterministicHashEmbedder``
  reference impl, hash + vector helpers (``content_hash``,
  ``cosine_similarity``, ``_hash_to_dim``, ``_hash_sign``,
  ``_normalise``), and the embedding-shape constants
  (``EMBEDDING_DIM``, ``FAKE_MODEL_ID``, ``_TOKEN_PATTERN``,
  ``_NOISE_STDDEV``, ``_UNIT_VECTOR_TOLERANCE``).
- ``storage.py`` — ``upsert_embedding`` + ``_existing_embedding`` write
  paths against ``LCMEmbedding``.
- ``semantic.py`` — ``SemanticHit``, ``semantic_search``,
  ``_fetch_embeddings``, ``_resolve_excerpts``, ``_excerpt_and_meta``,
  ``_SEMANTIC_CANDIDATE_CAP``. Pure semantic-cosine retrieval over the
  stored vectors.
- ``hybrid.py`` — ``LCMHybridSearchResult``, ``SearchMode``,
  ``ItemKind``, RRF blender (``reciprocal_rank_fusion``,
  ``_start_blend_from_lexical``, ``_start_blend_from_semantic``,
  ``_RRF_SMOOTHING``), and the public ``lcm_hybrid_search``
  orchestrator that fans out to semantic + lexical and folds them.

Public surface re-exported from ``embeddings/__init__.py``:
``EMBEDDING_DIM``, ``Embedder``, ``DeterministicHashEmbedder``,
``SemanticHit``, ``LCMHybridSearchResult``, ``SearchMode``,
``ItemKind``, ``content_hash``, ``cosine_similarity``,
``upsert_embedding``, ``semantic_search``,
``reciprocal_rank_fusion``, ``lcm_hybrid_search``. Verified by
grepping every ``from app.core.lcm.embeddings import …`` call site
(``app/core/lcm/evals.py``, ``tests/test_lcm_embeddings.py``) — every
imported name is in the list, no caller changes required.

Removed the ``backend/app/core/lcm/embeddings.py`` exemption from
``scripts/check-file-lines.mjs``. Only ``evals.py`` remains exempt from
the LCM trio; the comment was updated to reflect that.

Validation:
- ``node scripts/check-file-lines.mjs`` → OK with embeddings.py
  exemption removed (4 → 3 entries remain).
- ``cd backend && uv run mypy`` → 0 errors across 344 source files.
- ``cd backend && uv run ruff check .`` → clean.
- ``cd backend && uv run ruff format --check .`` → 345 files
  already formatted.
- ``pytest tests/test_lcm_embeddings.py tests/test_lcm_condensation.py``
  → 20 passed.
