+++ scripts/check-file-lines.mjs
@@ -90,15 +90,15 @@ const EXEMPT_PATH_FRAGMENTS = [
 	'backend/app/core/agent_loop/loop.py',
 	'backend/app/core/providers/claude_provider.py',
 	'backend/app/integrations/telegram/bot.py',
-	// LCM retrieval lab (PR #258) — two files still over budget. ``evals.py``
-	// is the largest (873 lines: harness + scenarios + answerer + retrievers
-	// all in one) and the natural split is into an ``evals/`` package
-	// (runner / seeding / answerer / retrievers). ``embeddings.py`` is
-	// storage + semantic + RRF; split semantic vs RRF blender.
-	// TODO(pawrrtal-lcm-split follow-up): land the two splits and
-	// remove this exemption.
+	// LCM retrieval lab (PR #258) — ``evals.py`` is still over budget
+	// (873 lines: harness + scenarios + answerer + retrievers all in
+	// one). The natural split is into an ``evals/`` package (runner /
+	// seeding / answerer / retrievers). ``embeddings.py`` was split
+	// into an ``embeddings/`` package (embedder / storage / semantic /
+	// hybrid) and no longer needs the exemption.
+	// TODO(pawrrtal-lcm-split follow-up): land the ``evals.py`` split
+	// and remove this exemption.
 	'backend/app/core/lcm/evals.py',
-	'backend/app/core/lcm/embeddings.py',
 ];
 
 /** Recursively yield every source file under `dir` that we should check. */
